% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_access.R
\name{access_reuse_chunk}
\alias{access_reuse_chunk}
\title{Primary logic for the "Access" section of the plan}
\usage{
access_reuse_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Primary logic for the "Access" section of the plan
}
\examples{
# Not run
access_reuse_chunk(input)
}
