% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{access_txt}
\alias{access_txt}
\title{Add markdown to text logic chunks for "Access" section}
\usage{
access_txt(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Add markdown to text logic chunks for "Access" section
}
\examples{
# Not run
access_txt(input)
}
