% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_cores.R
\name{check_files_tech}
\alias{check_files_tech}
\title{For a given core (research discipline), check whether its data types are
selected by the user in \code{input}}
\usage{
check_files_tech(input, core)
}
\arguments{
\item{input}{Shiny input}

\item{core}{core discipline, e.g., "genomics" or "em_tem"}
}
\value{
logical, is this core represented by the datatypes selected by the user?
}
\description{
For a given core (research discipline), check whether its data types are
selected by the user in \code{input}
}
\examples{
# Not run:
# Has the user selected antibody tech datatypes?
check_files_tech(input, "antibody_tech")
# Has the user selected genomics datatypes?
check_files_tech(input, "genomics")
}
