% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_datatype.R
\name{datatype_raw_by_core}
\alias{datatype_raw_by_core}
\title{Helps reactively update "raw_file_description" in server.R
Populates example text but allows user changes as long as the selected core
(core_datatype) is not changed.}
\usage{
datatype_raw_by_core(toggle_example_txt)
}
\arguments{
\item{toggle_example_txt}{frozen version of Shiny \code{input}}
}
\value{
character vector of possible values
}
\description{
Helps reactively update "raw_file_description" in server.R
Populates example text but allows user changes as long as the selected core
(core_datatype) is not changed.
}
\examples{
# Not run
# Pulls out the datatypes based on cores specified by the user
shiny::updateTextInput(session, "raw_file_description", value = datatype_raw_by_core(toggle_example_txt))
}
