% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{datatype_raw_file_description_options}
\alias{datatype_raw_file_description_options}
\title{#' Title
#'
#' @param input
#'
#' @return
#' @export
#'
#' @examples
processed_data_chunk <- function(input) {
return(
c(
"\emph{Our proposal will generate processed data of the following types and sizes:}  ",
paste0(
paste0(
"\strong{Technology:} We will process the <font color='OA799A'>",
if (input$raw_file_type == "") {
" ___ "
} else {
input$raw_file_type
},
"</font> using <font color='OA799A'>",
if (input$brief_pipeline_description == "") {
" ___ "
} else {
input$brief_pipeline_description
},
"</font>. "
),
paste0(
"\strong{File Type:} The data processing will result in <font color='OA799A'>",
if (input$processsed_file_description == "") {
" ___ "
} else {
input$processsed_file_description
},
"</font>. "
),
if (input$core_datatype \%in\% c("preclinical_img_microct")) {
paste0(
"The amount of data generated per timepoint <font color='OA799A'>",
if (input$avg_processed_file_size == "") {
" ___ "
} else {
input$avg_processed_file_size
},
"</font>. "
)
} else if (input$core_datatype == "proteomics") {
paste0(
"The amount of data generated will be <font color='OA799A'>",
if (input$avg_processed_file_size == "") {
" ___ "
} else {
input$avg_processed_file_size
},
"</font>. "
)
} else {
paste0(
"The amount of data generated per sample <font color='OA799A'>",
if (input$avg_processed_file_size == "") {
" ___ "
} else {
input$avg_processed_file_size
},
"</font>. "
)
},
paste0(
"\strong{Number of Files:} We anticipate generating <font color='OA799A'>",
if (input$total_processed_samples_files == "") {
" ___ "
} else {
input$total_processed_samples_files
},
"</font> for a total data volume of <font color='OA799A'>",
if (input$data_processed_volume == "") {
" ___ "
} else
(input$data_processed_volume)
,
"</font>. "
)
)
)
)
}}
\usage{
datatype_raw_file_description_options()
}
\description{
#' Title
#'
#' @param input
#'
#' @return
#' @export
#'
#' @examples
datatype_comment <- function(input) {
if (input$datatype_comment != "")
dt_comment <-
paste0("<font color='OA799A'>",
input$datatype_comment,
"</font>")
else {
dt_comment <- ""
}
return(dt_comment)
}
}
\details{
#' Title
#'
#' @param input
#'
#' @return
#' @export
#'
#' @examples
datatype_comment_summary <- function(input) {
if (input$datatype_comment_summary != "")
dt_comment_summary <-
paste0("<font color='OA799A'>",
input$datatype_comment_summary,
"</font>")
else {
dt_comment_summary <- ""
}
return(dt_comment_summary)
}
}
