% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{datatype_txt}
\alias{datatype_txt}
\title{Title}
\usage{
datatype_txt(
  input,
  raw_data_part,
  processed_data_part,
  shared_data_part,
  not_shared_data_part,
  metadata_part
)
}
\arguments{
\item{input}{}
}
\description{
Title
}
