% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_repository.R
\name{findable_chunk}
\alias{findable_chunk}
\title{Primary logic for the "Findable" section of the plan}
\usage{
findable_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Primary logic for the "Findable" section of the plan
}
\examples{
# Not run
findable_chunk(input)
}
