% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_cores.R
\name{flags_to_cores}
\alias{flags_to_cores}
\title{Get a vector of core names (not flags)}
\usage{
flags_to_cores(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
character vector of cores (research domains, e.g., "genomics",
"cell_img", etc.)
}
\description{
Not sure if this function is really needed. But gets the names of the cores
as indicated by the users instead of the dataframe of logical flags (above in
\code{determine_cores})
}
\examples{
# Not run
flags_to_cores(input)
}
