% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_access.R
\name{gru_hmb_options}
\alias{gru_hmb_options}
\title{Provide options for GRU vs HMB data use limitations}
\usage{
gru_hmb_options()
}
\value{
a named vector
}
\description{
Provide options for GRU vs HMB data use limitations
}
\examples{
gru_hmb_options()
}
