% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_metadata.R
\name{metadata_chunk}
\alias{metadata_chunk}
\title{Primary logic for the "Metadata" section of the plan}
\usage{
metadata_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Primary logic for the "Metadata" section of the plan
}
\examples{
# Not run
metadata_chunk(input)
}
