% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{oversight_txt}
\alias{oversight_txt}
\title{Add markdown to text logic chunks for "Oversight" section}
\usage{
oversight_txt(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Add markdown to text logic chunks for "Oversight" section
}
\examples{
# Not run
oversight_txt(input)
}
