% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{preservation_txt}
\alias{preservation_txt}
\title{Add markdown to text logic chunks for "Preservation" section}
\usage{
preservation_txt(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Add markdown to text logic chunks for "Preservation" section
}
\examples{
# Not run
preservation_txt(input)
}
