% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatype.R
\name{raw_processed_data_chunk}
\alias{raw_processed_data_chunk}
\title{#' Title
#'
#' @param input
#'
#' @return
#' @export
#'
#' @examples
raw_data_chunk <- function(input) {
return(c(
"\emph{Our proposal will generate raw data of the following types and sizes:}  ",}
\usage{
raw_processed_data_chunk(input)
}
\arguments{
\item{input}{}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{paste0(
  paste0(if (input$core_datatype \%in\% c(
    "antibody_tech",
    "cellular_imaging",
    "em_tem",
    "em_sem",
    "em_cryoem",
    "eh_aperio",
    "eh_polaris",
    "eh_vectra",
    "flow_cytometry",
    "genomics",
    "immune",
    "large_animal",
    "preclinical_img_ivis",
    "preclinical_img_microct",
    "preclinical_img_mri",
    "proteomics",
    "small_animal",
    "therapeutic"
  )) \{
    "**Technology:** We will collect <font color='OA799A'>"
  \} else \{
    "**Technology:** We will generate data using <font color='OA799A'>"
  \},
  if (length(input$technology_description) == 0) \{" ___ "\} else \{paste(input$technology_description, collapse = ", ")\},
  "</font>. "),
  paste0(
    "**File Type:** Data for this study will generate <font color='OA799A'>",
    if (input$raw_file_description == "") \{
      " ___ "
    \} else \{
      input$raw_file_description
    \},
    "</font>. "
  ),
}\if{html}{\out{</div>}}
}
\details{

}
