% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_reference.R
\name{reference_chunk}
\alias{reference_chunk}
\title{Primary logic for the "Notes" section.}
\usage{
reference_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
These urls are generated for the user's reference and should not be
included in the final plan, as links to websites are not allowed in the
submitted plan.
}
\details{
This text appears below the rendered HTML of the plan itself and is not
downloaded with the plan.
}
\examples{
# Not run
reference_chunk(input)
}
