% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_last_comma}
\alias{remove_last_comma}
\title{Fix list grammar}
\usage{
remove_last_comma(string_vect, conj = "and")
}
\arguments{
\item{string_vect}{}
}
\description{
Goal of this function is to take a string like a, b, c and return a, b, and c.
}
