% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{shiny_ui}
\alias{shiny_ui}
\title{Create the UI side for the Shiny app}
\usage{
shiny_ui()
}
\value{
UI side of Shiny for use in shinyApp(). Includes some static text.
}
\description{
Create the UI side for the Shiny app
}
\examples{
# Not run
shinyApp(shiny_ui, shiny_server)
}
