% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_standards.R
\name{standards_chunk}
\alias{standards_chunk}
\title{Primary logic for the "Standards" section of the plan}
\usage{
standards_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Primary logic for the "Standards" section of the plan
}
\examples{
# Not run
standards_chunk(input)
}
