% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{standards_txt}
\alias{standards_txt}
\title{Add markdown to text logic chunks for "Standards" section}
\usage{
standards_txt(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Add markdown to text logic chunks for "Standards" section
}
\examples{
# Not run
standards_txt(input)
}
