% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{tools_txt}
\alias{tools_txt}
\title{Add markdown to text logic chunks for "Tools" section}
\usage{
tools_txt(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Add markdown to text logic chunks for "Tools" section
}
\examples{
# Not run
tools_txt(input)
}
