% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_access.R
\name{control_chunk}
\alias{control_chunk}
\title{Primary logic for the "Access Control" section of the plan}
\usage{
control_chunk(input)
}
\arguments{
\item{input}{Shiny input}
}
\value{
a character vector containing the text for this section
}
\description{
Primary logic for the "Access Control" section of the plan
}
\examples{
# Not run
control_chunk(input)
}
