% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_datatype.R
\name{datatype_raw_by_core}
\alias{datatype_raw_by_core}
\title{Title
Reactively updates "raw_file_description" in server.R}
\usage{
datatype_raw_by_core(toggle_example_txt)
}
\arguments{
\item{toggle_example_txt}{}
}
\description{
Title
Reactively updates "raw_file_description" in server.R
}
