% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{datatype_txt}
\alias{datatype_txt}
\title{Title}
\usage{
datatype_txt(
  technology_description,
  raw_file_description,
  avg_file_size,
  total_samples_files,
  data_volume,
  raw_file_type,
  brief_pipeline_description,
  processsed_file_description,
  avg_processed_file_size,
  total_processed_samples_files,
  data_processed_volume,
  datatype_comment,
  shared_technologies,
  notshared,
  metadata_part
)
}
\arguments{
\item{data_processed_volume}{}
}
\description{
Title
}
