% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{datatype_txt}
\alias{datatype_txt}
\title{Title}
\usage{
datatype_txt(
  raw_part,
  total_samples_files,
  data_volume,
  processed_part,
  avg_processed_file_size,
  total_processed_samples_files,
  data_processed_volume,
  datatype_comment,
  shared_technologies,
  notshared,
  metadata_part
)
}
\arguments{
\item{metadata_part}{}
}
\description{
Title
}
