% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{preservation_txt}
\alias{preservation_txt}
\title{Title}
\usage{
preservation_txt(repository_part, fair_description, duration_description)
}
\arguments{
\item{duration_description}{}
}
\description{
Title
}
